Deprecate async hooks and provide with* replacements
The default ESLint rule for React hooks errors on calling functions like  `useNode` from async functions
(server components). It should not be necessary to configure ESLint rules to use Zebra, so we rather change
the names of our functions now. Because they are not _really_ hooks I settled for the name `with*` to
differentiate from the actual data loaders.